+++ packages/components/src/utils/undefinedCoalesce.ts
@@ -24,36 +24,6 @@
 
  */
 
-import React, { FC, ReactNode } from 'react'
-import styled from 'styled-components'
-import { CompatibleHTMLProps } from '@looker/design-tokens'
-import { Space, SpaceHelperProps } from '../../Layout/Space'
-
-export interface DialogFooterProps
-  extends CompatibleHTMLProps<HTMLDivElement>,
-    SpaceHelperProps {
-  /**
-   * Secondary content to place in the footer
-   */
-  secondary?: ReactNode
-}
-
-export const DialogFooterLayout: FC<DialogFooterProps> = ({
-  children,
-  secondary,
-  ...props
-}) => {
-  return (
-    <Space as="footer" reverse between {...props}>
-      <Space reverse>{children}</Space>
-      {secondary && <Space>{secondary}</Space>}
-    </Space>
-  )
-}
-
-export const DialogFooter = styled(DialogFooterLayout)``
-
-DialogFooter.defaultProps = {
-  px: 'xlarge',
-  py: 'large',
+export const undefinedCoalesce = (arr: any[]) => {
+  return arr.find((element) => element !== undefined)
 }