+++ CHANGELOG.md
@@ -9,19 +9,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Storybook is now deployed to `https://components.looker.com/storybook`
+- `undefinedCoalesce` util function
 - `Link` now supports explicit `underline` & `keyColor` properties
 - `FactCheck` icon
 
 ### Changed
 
+- `Accordion`
+  - Updated padding props to affect `AccordionDisclosure` and `AccordionContent` containers rather than parent `Accordion` container
+- `AccordionDisclosure`
+  - Removed default `height: 100%`
+- `Tree` and `TreeItem`
+  - Updated docs and test suite
 - `theme.zIndexFloor` now defaults to `1` instead of `undefined` (this helps with compatibility due to Firefox's stacking order eccentricities)
 - `Tooltip` now renders in a `Portal`
+- `TabList` now supports distribute prop
+- `MessageBar`
+  - Sets `canDismiss` to true by default
+  - Adds new `visible` prop to toggle display externally
 
 ### Fixed
 
 - `Select`/`SelectMulti` keyboard navigation when filtering and going from > 100 to < 100 options
 - `SelectMulti` with `freeInput` not saving input value on tab key
 - `SelectMulti` list not closing on blur
+
+### Removed
+
+- Removed `Modal*` aliases to `Dialog*`
 
 ## [0.9.5] - 2020-07-01